+++ src/Twilio/Rest/Video/V1/RoomList.php
@@ -55,6 +55,8 @@ class RoomList extends ListResource {
             'RecordingRules' => Serialize::jsonObject($options['recordingRules']),
             'AudioOnly' => Serialize::booleanToString($options['audioOnly']),
             'MaxParticipantDuration' => $options['maxParticipantDuration'],
+            'EmptyRoomTimeout' => $options['emptyRoomTimeout'],
+            'UnusedRoomTimeout' => $options['unusedRoomTimeout'],
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);