+++ tests/Twilio/Integration/Video/V1/RoomTest.php
@@ -55,6 +55,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "audio_only": false,
                 "media_region": "us1",
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -110,6 +112,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "media_region": "us1",
                 "audio_only": false,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -151,6 +155,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "media_region": "us1",
                 "audio_only": false,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -192,6 +198,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "media_region": "us1",
                 "audio_only": false,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -231,6 +239,8 @@ class RoomTest extends HolodeckTestCase {
                 "video_codecs": [],
                 "media_region": "us1",
                 "audio_only": true,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -272,6 +282,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "media_region": "us1",
                 "audio_only": false,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -313,6 +325,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "media_region": "us1",
                 "audio_only": false,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -352,6 +366,8 @@ class RoomTest extends HolodeckTestCase {
                 "video_codecs": [],
                 "media_region": "us1",
                 "audio_only": true,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -434,6 +450,8 @@ class RoomTest extends HolodeckTestCase {
                         ],
                         "media_region": "us1",
                         "audio_only": false,
+                        "empty_room_timeout": 5,
+                        "unused_room_timeout": 5,
                         "url": "https://video.twilio.com/v1/Rooms/RM4070b618362c1682b2385b1f9982833c",
                         "links": {
                             "participants": "https://video.twilio.com/v1/Rooms/RM4070b618362c1682b2385b1f9982833c/Participants",
@@ -502,6 +520,8 @@ class RoomTest extends HolodeckTestCase {
                 ],
                 "media_region": "us1",
                 "audio_only": false,
+                "empty_room_timeout": 5,
+                "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
                 "duration": 10,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",